+++ python/src/trezorlib/cli/debug.py
@@ -16,59 +16,12 @@
 
 import click
 
-from .. import debuglink, mapping, messages, protobuf
-from ..messages import DebugLinkShowTextStyle as S
-from . import with_client
+from .. import mapping, messages, protobuf
 
 
 @click.group(name="debug")
 def cli():
     """Miscellaneous debug features."""
-
-
-STYLES = {
-    "@@NORMAL": S.NORMAL,
-    "@@BOLD": S.BOLD,
-    "@@MONO": S.MONO,
-    "@@BR": S.BR,
-    "@@BR_HALF": S.BR_HALF,
-}
-
-
-@cli.command()
-@click.option("-i", "--icon", help="Header icon name")
-@click.option("-c", "--color", help="Header icon color")
-@click.option("-h", "--header", help="Header text", default="Showing text")
-@click.argument("body")
-@with_client
-def show_text(client, icon, color, header, body):
-    """Show text on Trezor display.
-
-    For usage instructions, see:
-    https://github.com/trezor/trezor-firmware/blob/master/docs/python/show-text.md
-    """
-    body = body.split()
-    body_text = []
-    words = []
-
-    def _flush():
-        if words:
-            body_text.append((None, " ".join(words)))
-        words.clear()
-
-    for word in body:
-        if word in STYLES:
-            _flush()
-            body_text.append((STYLES[word], None))
-        elif word.startswith("%%"):
-            _flush()
-            body_text.append((S.SET_COLOR, word[2:]))
-        else:
-            words.append(word)
-
-    _flush()
-
-    return debuglink.show_text(client, header, body_text, icon=icon, icon_color=color)
 
 
 @cli.command()