+++ python/src/trezorlib/debuglink.py
@@ -664,18 +664,3 @@ def self_test(client):
             payload=b"\x00\xFF\x55\xAA\x66\x99\x33\xCCABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!\x00\xFF\x55\xAA\x66\x99\x33\xCC"
         )
     )
-
-
-@expect(messages.Success, field="message")
-def show_text(client, header_text, body_text, icon=None, icon_color=None):
-    body_text = [
-        messages.DebugLinkShowTextItem(style=style, content=content)
-        for style, content in body_text
-    ]
-    msg = messages.DebugLinkShowText(
-        header_text=header_text,
-        body_text=body_text,
-        header_icon=icon,
-        icon_color=icon_color,
-    )
-    return client.call(msg)